feat: add Raft data structure definitions
Add the bare definitions for the Raft server state and RPC messages.
Currently these just store the raw data for the algorithm as specified
in Figure 2 of the Raft paper.
The `RaftServer` struct also includes some stub methods for receiving
and responding to RPC requests.

Two tests are included, verifying the startup state of the Raft server
and the `RaftServer::set_state` function.