feat: implement RaftServer::recv_append_entries
Implement the `RaftServer::recv_append_entries` function which accepts
an AppendEntries RPC message and appends the new entries in the message
to the log, provided all necessary criteria are met.
Add tests for all receiver implementation conditions for
`recv_append_entries` as indicated in Figure 2 of the Raft paper.

Change the `LogEntry` type to be a proper struct type to allow
implementation of methods.
Add a `LogEntry::term` function to get the term of a given log entry.

Implement the `set_current_term` and `set_commit_index` setter functions
for `RaftServer`.