fix: ensure `voted_for` field is set after RequestVote succeeds
Add missing `voted_for` assignment the the RequestVote RPC handler to
ensure that the candidate being voted for is updated when the RPC
executes successfully.

Add `RaftServer::set_voted_for` function.

Add an `assert!` to ensure that `msg.term` is never greater than
`self.current_term` in the RequestVote RPC handler.

Modify the log up-to-date checking logic to use `cmp` with a match
pattern, rather than nested if-else.

Add `last_term` helper function to get the term of the last entry in the
log.

Add extra assertions to RequestVote RPC handler tests to validate
invariants.