feat: implement `RaftServer::update_term`
Implement `RaftServer::update_term` function which updates a server's
term in the event than an RPC call received which specifies a newer term
than that of the server.

Implement `Message` trait on all RPC messages and responses; `Message`
provides a `term` function which allows functions such as `update_term`
to get term information in a message-agnostic manner.