feat: implement `RaftServer::receive`
Implement the `RaftServer::receive` function which takes in a `Message`
and either: updates the server's term if the RPC has a newer term, drops
the response if the term is stale, or processes the RPC, returning a
response as necessary.
Note that the function returns a `ReceiveResult` enum indicating what
occured as part of the function call. It is up to the callee to discard
the message once it has been marked as stale or processed.
It is also up to the callee to send the response over the network, if one has
been generated.

A `Message` enum was added for encapsulation of the 4 main RPC message
types.
This enum has a `term` function which may be called to get the `term`
the message was issued in.
The `Message` trait was obsoleted in favour of this enum.

The `RaftServer::recv_` functions have been updated to take references
to messages instead of ingesting the messages.
This is an unfortunate consequence of the fact that the
`RaftServer::receive` function calls these methods, and
`RaftServer::receive` itself does not always consume the message (e.g.
if the server's term must first be updated).

Add tests for the `RaftServer::receive` function to ensure that all
types of messages are dropped/processed/etc accordingly.