fix: set commit_index to min(leader_commit, idx_last_entry)
Fix the logic for setting the `commit_index` when receiving an
AppendEntries RPC call. Previously, we simply set the `commit_index` to
`leader_commit`, per the TLA+ specification, but according to the paper
this is incorrect, and it must be set to `min(leader_commit,
idx_last_entry)` instead.

Additionally, in the future it should be determined if we should follow
the specification's procedure and only update the `commit_index` after
handling a heartbeat or a repeated successful RPC call (i.e., don't
update the `commit_index` when handling a conflict).