feat: add basic FFI functionality
Add `ffi` module to `castaway`.
Add a `CVec` utility type for representing a safe abstraction over a C
array.
Add `raft_server_new` and `raft_server_free` FFI functions to allow
users to instantiate a `RaftServer` and then free it when done.

Add a `castaway.py` and `castaway_cffi.py`.
`castaway_cffi.py` uses the Python `cffi` module to auto-generate the
necessary bindings to call libcastaway functions from Python.
`castaway.py` provides an example (toy) implementation of Raft on top of
libcastaway, in Python.
Currently the script just initializes and frees a set of RaftServers.
The servers are not properly configured for use yet.

Add the `__pycache__` directory to the gitignore.
Add `libc` as a Rust dependency.